fix-auto restart

+++ start.js
@@ -1,28 +1,32 @@
 import fs from "fs"
-import path from 'path';
-import cluster from 'cluster'
+import path from "path"
+import cluster from "cluster"
 import "./setting.js"
-
 
 let isRunning = false
 
 async function start(file) {
     if (isRunning) return
     isRunning = true
+
     cluster.setupMaster({ exec: path.join(__dirname, file), args: [] })
     let p = cluster.fork()
-    p.on(`message`, async function (data) {
-        switch (data) {
-            case `restart`: {
-                p.process.kill()
-                isRunning = false
-                await start(file)
-            }
+
+    p.on("message", async (data) => {
+        if (data === "restart") {
+            p.process.kill()
+            isRunning = false
+            await start(file)
         }
     })
-    p.on(`exit`, async function (code) {
-        start(file)
-        console.log(`sistem mati dengan kode: ${code}`)
+
+    p.on("exit", async (code) => {
+        console.log("sistem mati dengan kode:", code)
+        if (!code || parseInt(code) === 0) return
+        console.log("start ulang")
+        isRunning = false
+        await start(file)
     })
 }
-start(`connect.js`)
+
+start("connect.js")
